+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release History
-# 1.3.9 (Unreleased)
+
+# 1.3.9 (2026-09-09)
 ### Features Added
 - - Update OpenTelemetry dependencies to latest versions, bump `langchain-core` minimum version to address S360, and support the new `httpx2` entry point exposed by `opentelemetry-instrumentation-httpx`.
   ([#254](https://github.com/microsoft/opentelemetry-distro-python/pull/254))